pip_requirements.txt
# Runtime dependencies
numpy  # ComEd price statistics (median, percentile, regression)
pywemo  # WeMo smart plug control for battery charge/discharge relay
pyyaml  # YAML configuration file parsing
requests  # HTTP client for EP Cube and ComEd APIs
scipy  # ComEd price trend prediction (linear regression)
